+++ src/functional/thermo_controll.h
@@ -42,9 +42,9 @@ void driveHeaterByEnergyBalance(HeatState& heatState, float currentTemp, float s
 
 float caculateCorrection(HeatState& heatState, float currentTemp, float setPoint);
 
-float doPIDAdjust(float targetTemp, PID& pidController, const SensorState& currentState, HeatState& heatState);
+double doPIDAdjust(float targetTemp, PID& pidController, const SensorState& currentState, HeatState& heatState);
 
-float doPIDAdjustWithLimit(float targetTemp, float downLimit, float upperLimit, PID& pidController, const SensorState& currentState, HeatState& heatState);
+double doPIDAdjustWithLimit(float targetTemp, float downLimit, float upperLimit, PID& pidController, const SensorState& currentState, HeatState& heatState);
 
 void pulseHeaters(HeatState& heatState, unsigned long powerAdjustCycle);
